atualização do projeto com a implementação da camada MVC ( falta efetivar o Singleton para acesso ao banco de dados, assim como falta implementar todo o frontend e o Singleton para acessar os Controls dentro do javascript ).
