+++ Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/DataBase/repositories/personRepository.java
@@ -1,0 +1,7 @@
+package br.unoeste.fipp.socialcare.DataBase.repositories;
+
+import br.unoeste.fipp.socialcare.DataBase.entities.Person;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface personRepository extends JpaRepository<Person, Long> {
+}
