+++ Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/DataBase/repositories/storageRepository.java
@@ -1,0 +1,8 @@
+package br.unoeste.fipp.socialcare.DataBase.repositories;
+
+import br.unoeste.fipp.socialcare.DataBase.entities.Storage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface storageRepository extends JpaRepository<Storage, Long> {
+
+}
