+++ Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/DataBase/repositories/userRepository.java
@@ -1,0 +1,7 @@
+package br.unoeste.fipp.socialcare.DataBase.repositories;
+
+import br.unoeste.fipp.socialcare.DataBase.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface userRepository extends JpaRepository<User, Long> {
+}
