+++ Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/RestControls/userRestControl.java
@@ -1,0 +1,16 @@
+package br.unoeste.fipp.socialcare.RestControls;
+
+import org.springframework.web.bind.annotation.*;
+
+
+@RestController
+@RequestMapping(value="apis/citizen/")
+public class userRestControl {
+
+    @GetMapping(value="connection-test")
+    public String connectionTest(){
+        return "connected";
+    }
+
+
+}
