+++ Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/Singleton/SingletonAdmin.java
@@ -1,0 +1,16 @@
+package br.unoeste.fipp.socialcare.Singleton;
+
+public class SingletonAdmin {
+
+    private static SingletonAdmin instance;
+
+    private SingletonAdmin() {}
+
+    public static SingletonAdmin getInstance()
+    {
+        if (instance==null)
+            instance = new SingletonAdmin();
+        return (instance);
+    }
+
+}
