+++ Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/View/frontend/index.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html lang="pt-BR">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Lar dos meninos</title>
+    <link rel="stylesheet" href="style.css">
+</head>
+<body>
+    <header>
+        <h1>Lar dos meninos - Presidente Prudente</h1>
+    </header>
+    <section>
+        <h2>Quem Somos</h2>
+        <p>Somos uma ONG dedicada a ajudar os mais necessitados e promover a solidariedade na comunidade.</p>
+    </section>
+    <section>
+        <h2>Operações</h2>
+        <p>Aqui você pode encontrar informações sobre as atividades e projetos que realizamos para ajudar quem mais precisa.</p>
+    </section>
+    <section>
+        <h2>Como Ajudar</h2>
+        <p>Se você deseja contribuir com nossa causa, entre em contato para saber como ajudar financeiramente, como voluntário ou doando alimentos e roupas.</p>
+    </section>
+    <footer>
+        <p>&copy; 2021 ONG Solidariedade. Todos os direitos reservados.</p>
+    </footer>
+</body>
+</html>
