--- Projeto_SocialCare/backend/SocialCare/src/main/java/br/unoeste/fipp/socialcare/db/entities/CategoryProduct.java
@@ -1,42 +1,0 @@
-package br.unoeste.fipp.socialcare.db.entities;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name="categoria")
-public class CategoryProduct {
-
-    //Atributos proprios da tabela de categoria
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="cat_id")
-    private Long id;
-    @Column(name="cat_nome")
-    private String nome;
-
-
-    public CategoryProduct() {
-        this(0L,"");
-    }
-
-    public CategoryProduct(Long id, String nome) {
-        this.id = id;
-        this.nome = nome;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-}
